Feature : [#5] refactor package structure
