--- src/main/java/com/abcmart/shoestore/domain/CashPayment.java
@@ -1,22 +1,0 @@
-package com.abcmart.shoestore.domain;
-
-import com.abcmart.shoestore.tool.PaymentType;
-import java.math.BigDecimal;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CashPayment extends Payment {
-
-
-    private CashPayment(BigDecimal paidAmount) {
-
-        super(PaymentType.CASH, paidAmount);
-    }
-
-    public static CashPayment payInCash(BigDecimal paidAmount) {
-
-        return new CashPayment(paidAmount);
-    }
-}
