+++ src/main/java/com/abcmart/shoestore/order/domain/OrderStatus.java
@@ -1,0 +1,18 @@
+package com.abcmart.shoestore.order.domain;
+
+public enum OrderStatus {
+
+    NORMAL,
+    CANCEL
+    ;
+
+    public boolean isNormal() {
+
+        return this == NORMAL;
+    }
+
+    public boolean isCancel() {
+
+        return this == CANCEL;
+    }
+}
