+++ src/main/java/com/abcmart/shoestore/order/dto/OrderDto.java
@@ -1,0 +1,35 @@
+package com.abcmart.shoestore.order.dto;
+
+import com.abcmart.shoestore.order.domain.Order;
+import com.abcmart.shoestore.payment.dto.PaymentDto;
+import com.abcmart.shoestore.order.domain.OrderStatus;
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class OrderDto {
+
+    private final Long orderNo;
+    private final OrderStatus status;
+    private final List<OrderDetailDto> details;
+    private final List<PaymentDto> payments;
+
+    private OrderDto(Long orderNo, OrderStatus status, List<OrderDetailDto> details,
+        List<PaymentDto> payments) {
+
+        this.orderNo = orderNo;
+        this.status = status;
+        this.details = details;
+        this.payments = payments;
+    }
+
+    public static OrderDto from(Order order) {
+
+        return new OrderDto(
+            order.getOrderNo(),
+            order.getStatus(),
+            order.getDetails().stream().map(OrderDetailDto::from).toList(),
+            order.getPayments().values().stream().map(PaymentDto::from).toList()
+        );
+    }
+}
