+++ src/main/java/com/abcmart/shoestore/payment/domain/CreditCardType.java
@@ -1,0 +1,15 @@
+package com.abcmart.shoestore.payment.domain;
+
+public enum CreditCardType {
+
+    SHINHAN,
+    SAMSUNG,
+    KB_KOOKMIN,
+    HYUNDAI,
+    LOTTE,
+    WOORI,
+    HANA,
+    BC
+    ;
+
+}
