+++ src/main/java/com/abcmart/shoestore/payment/domain/PaymentType.java
@@ -1,0 +1,18 @@
+package com.abcmart.shoestore.payment.domain;
+
+public enum PaymentType {
+
+    CASH,
+    CREDIT_CARD
+    ;
+
+    public boolean isCash() {
+
+        return this == CASH;
+    }
+
+    public boolean isCreditCard() {
+
+        return this == CREDIT_CARD;
+    }
+}
