+++ src/main/java/com/abcmart/shoestore/payment/repository/PaymentRepository.java
@@ -1,0 +1,15 @@
+package com.abcmart.shoestore.payment.repository;
+
+import com.abcmart.shoestore.payment.domain.CardPayment;
+import com.abcmart.shoestore.payment.domain.Payment;
+import java.util.List;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PaymentRepository {
+
+    Payment save(Payment payment);
+
+    List<CardPayment> findAllCreditCardPayments();
+
+}
