--- src/main/java/com/abcmart/shoestore/repository/OrderRepository.java
@@ -1,16 +1,0 @@
-package com.abcmart.shoestore.repository;
-
-import com.abcmart.shoestore.domain.Order;
-import com.abcmart.shoestore.domain.OrderDetail;
-import java.util.List;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderRepository {
-
-    Order save(Order order);
-
-    Order findByOrderNo(Long orderNo);
-
-    List<OrderDetail> findAllNormalStatusOrderDetails();
-}
