--- src/main/java/com/abcmart/shoestore/repository/PaymentRepository.java
@@ -1,15 +1,0 @@
-package com.abcmart.shoestore.repository;
-
-import com.abcmart.shoestore.domain.CardPayment;
-import com.abcmart.shoestore.domain.Payment;
-import java.util.List;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PaymentRepository {
-
-    Payment save(Payment payment);
-
-    List<CardPayment> findAllCreditCardPayments();
-
-}
