--- src/main/java/com/abcmart/shoestore/tool/CreditCardType.java
@@ -1,15 +1,0 @@
-package com.abcmart.shoestore.tool;
-
-public enum CreditCardType {
-
-    SHINHAN,
-    SAMSUNG,
-    KB_KOOKMIN,
-    HYUNDAI,
-    LOTTE,
-    WOORI,
-    HANA,
-    BC
-    ;
-
-}
