--- src/main/java/com/abcmart/shoestore/tool/OrderStatus.java
@@ -1,18 +1,0 @@
-package com.abcmart.shoestore.tool;
-
-public enum OrderStatus {
-
-    NORMAL,
-    CANCEL
-    ;
-
-    public boolean isNormal() {
-
-        return this == NORMAL;
-    }
-
-    public boolean isCancel() {
-
-        return this == CANCEL;
-    }
-}
