+++ src/test/java/com/abcmart/shoestore/application/AdminServiceTest.java
@@ -7,17 +7,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
 
-import com.abcmart.shoestore.application.response.ShoeSaleAmountResponse;
-import com.abcmart.shoestore.application.response.ShoeSaleAmountResponse.CreditCardSaleAmountResponse;
-import com.abcmart.shoestore.application.response.ShoeSaleCountResponse;
-import com.abcmart.shoestore.application.response.ShoeSaleCountResponse.SoldShoe;
-import com.abcmart.shoestore.domain.CardPayment;
-import com.abcmart.shoestore.domain.OrderDetail;
-import com.abcmart.shoestore.domain.Shoe;
-import com.abcmart.shoestore.repository.OrderRepository;
-import com.abcmart.shoestore.repository.PaymentRepository;
-import com.abcmart.shoestore.repository.ShoeRepository;
-import com.abcmart.shoestore.tool.CreditCardType;
+import com.abcmart.shoestore.admin.application.AdminService;
+import com.abcmart.shoestore.admin.application.response.ShoeSaleAmountResponse;
+import com.abcmart.shoestore.admin.application.response.ShoeSaleAmountResponse.CreditCardSaleAmountResponse;
+import com.abcmart.shoestore.admin.application.response.ShoeSaleCountResponse;
+import com.abcmart.shoestore.admin.application.response.ShoeSaleCountResponse.SoldShoe;
+import com.abcmart.shoestore.payment.domain.CardPayment;
+import com.abcmart.shoestore.order.domain.OrderDetail;
+import com.abcmart.shoestore.shoe.domain.Shoe;
+import com.abcmart.shoestore.order.repository.OrderRepository;
+import com.abcmart.shoestore.payment.repository.PaymentRepository;
+import com.abcmart.shoestore.shoe.repository.ShoeRepository;
+import com.abcmart.shoestore.payment.domain.CreditCardType;
 import java.math.BigDecimal;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
