+++ src/test/java/com/abcmart/shoestore/application/OrderServiceTest.java
@@ -11,21 +11,22 @@ import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 
-import com.abcmart.shoestore.application.request.CreateOrderRequest;
-import com.abcmart.shoestore.application.request.CreateOrderRequest.CreateOrderDetailRequest;
-import com.abcmart.shoestore.application.request.CreateOrderRequest.CreatePaymentRequest;
-import com.abcmart.shoestore.domain.Order;
-import com.abcmart.shoestore.domain.OrderDetail;
-import com.abcmart.shoestore.domain.Payment;
-import com.abcmart.shoestore.domain.Shoe;
-import com.abcmart.shoestore.dto.OrderDetailDto;
-import com.abcmart.shoestore.dto.OrderDto;
-import com.abcmart.shoestore.dto.PaymentDto;
-import com.abcmart.shoestore.repository.OrderRepository;
-import com.abcmart.shoestore.repository.ShoeRepository;
-import com.abcmart.shoestore.tool.CreditCardType;
-import com.abcmart.shoestore.tool.OrderStatus;
-import com.abcmart.shoestore.tool.PaymentType;
+import com.abcmart.shoestore.order.application.request.CreateOrderRequest;
+import com.abcmart.shoestore.order.application.request.CreateOrderRequest.CreateOrderDetailRequest;
+import com.abcmart.shoestore.order.application.request.CreateOrderRequest.CreatePaymentRequest;
+import com.abcmart.shoestore.order.domain.Order;
+import com.abcmart.shoestore.order.domain.OrderDetail;
+import com.abcmart.shoestore.payment.domain.Payment;
+import com.abcmart.shoestore.shoe.domain.Shoe;
+import com.abcmart.shoestore.order.application.OrderService;
+import com.abcmart.shoestore.order.dto.OrderDetailDto;
+import com.abcmart.shoestore.order.dto.OrderDto;
+import com.abcmart.shoestore.payment.dto.PaymentDto;
+import com.abcmart.shoestore.order.repository.OrderRepository;
+import com.abcmart.shoestore.shoe.repository.ShoeRepository;
+import com.abcmart.shoestore.payment.domain.CreditCardType;
+import com.abcmart.shoestore.order.domain.OrderStatus;
+import com.abcmart.shoestore.payment.domain.PaymentType;
 import java.math.BigDecimal;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
